# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
# Default security: every operation accepts a session cookie. Public
# operations (e.g. /menu/public, /auth/phone/send-otp) override with
# `security: []` to declare they do not require auth. Admin operations
# may additionally accept `adminToken`.
security:
  - sessionCookie: []
...
paths:
...
  /orders/finalize:
    post:
      operationId: finalizeOrder
      tags: [orders]
      summary: |
        Atomically persist an order, redeem credits, and award referrals
        in one transaction. Server-managed idempotency is REQUIRED — see
        the `Idempotency-Key` header.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinalizeOrderRequest" }
      responses:
        "200":
          description: |
            Order persisted (or cached response replayed for a retry of
            the same key + body). Replays carry `Idempotent-Replay: true`.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderFinalizeResult" }
        "400": { description: Invalid payload or missing/invalid Idempotency-Key }
        "401": { description: Unauthorized }
        "409": { description: Idempotency-Key reused with a different body }
...
components:
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9._\\-:]{8,128}$"
      description: |
        Required server-managed idempotency token. 8–128 characters,
        url-safe alphabet `[A-Za-z0-9._-:]`. Reuse the same value for
        every retry of the SAME submit attempt; pick a fresh value
        (e.g. crypto.randomUUID()) for each NEW user-initiated submit.
  securitySchemes:
    # Phone-OTP session — server reads the `sid` cookie set by
    # /auth/phone/verify-otp.
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: |
        Phone-OTP session id. Set automatically by the server on
        /auth/phone/verify-otp. Sent on every same-origin request as
        an httpOnly cookie. Routes annotated "(auth required)" require
        a valid session.
    # Operator / RD admin static-token gate. Sent in the
    # `x-admin-token` request header for ops & catalog endpoints.
    adminToken:
      type: apiKey
      in: header
      name: x-admin-token
      description: |
        Operator-only static token for admin / catalog routes. Falls
        back to OPS_USER_IDS / CATALOG_USER_IDS allowlist matching
        when the caller has a valid session cookie.
  schemas:
    FinalizeOrderRequest:
      type: object
      required: [orderId, items]
      additionalProperties: true
      properties:
        orderId: { type: string, description: Client-generated order id }
        items:
          type: array
          items:
            type: object
            required: [id, name, qty, price]
            properties:
              id: { type: integer }
              name: { type: string }
              qty: { type: integer, minimum: 1 }
              price: { type: integer, description: Per-unit price in paise }
        applyCreditsPaise: { type: integer, nullable: true }
        scheduledFor: { type: string, format: date-time, nullable: true }
        bundleSlugs: { type: array, items: { type: string } }
        fulfillmentType: { type: string, enum: [delivery, pickup] }
        deliverySlotId: { type: integer, nullable: true }
        pickupLocationId: { type: integer, nullable: true }
        ecoPackagingOptIn: { type: boolean }
        deliveryInstructions: { type: string, nullable: true }
    OrderFinalizeResult:
      type: object
      required: [ok]
      additionalProperties: true
      properties:
        ok: { type: boolean }
        serverOrderId: { type: integer }
        finalTotalPaise: { type: integer }
        creditsRedeemedPaise: { type: integer }
        firstOrderDiscountPaise: { type: integer }
        referralAwarded: { type: boolean }
...