# lib/api-zod/src/generated/types/orderFinalizeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderFinalizeResult {
  ok: boolean;
  serverOrderId?: number;
  finalTotalPaise?: number;
  creditsRedeemedPaise?: number;
  firstOrderDiscountPaise?: number;
  referralAwarded?: boolean;
  [key: string]: unknown;
}